requirements.txt
fastapi
uvicorn
alembic
pydantic-settings